// src/enums/ServerStats.enum.ts
export enum ServerStatsPrefix {
    CountMembersAll = "%COUNT_MEMBERS_ALL%",
    CountMembersUsers = "%COUNT_MEMBERS_USERS%",
    CountMembersBots = "%COUNT_MEMBERS_BOTS%",
    CountMembersOnline = "%COUNT_MEMBERS_ONLINE%",
    CountMembersDnd = "%COUNT_MEMBERS_DND%",
    CountMembersIdle = "%COUNT_MEMBERS_IDLE%",
    CountMembersOffline = "%COUNT_MEMBERS_OFFLINE%",
    CountChannelsVoice = "%COUNT_CHANNELS_VOICE%",
    CountChannelsText = "%COUNT_CHANNELS_TEXT%"
}